Fix three campaign-halting bugs; add fail-fast pre-flight for deep baselines
The campaign crashed three times. Each crash is fixed at its source, and each
fix ships with a test that fails against the old code.

1. Long-dtype crash in the classical baselines
   panel.py builds counts as int64 (correct -- counts are integers) and the
   dataset sliced the window through unchanged, so historical-average hit
   "mean(): could not infer output dtype ... Got: Long". The main model never
   hit it because it reaches counts via log1p, which promotes on its own.
   Fixed once at the dataset boundary rather than sprinkling .float() at each
   call site. Provably neutral for the main path: log1p(long) and
   log1p(float32) are bit-identical, so no published number can move.

2. STZINB-GNN diffused across the wrong axis
   'ij,bsjh->bsih' binds j to the WINDOW axis, so a spatial adjacency was
   contracted against time. It crashed at S=77/W=52 after ~1000s x 3 seeds.
   The worse property is silent: at S == W both forms are shape-legal and the
   buggy one trains and reports plausible, wrong numbers. Now
   'ij,bjwh->biwh', checked against an explicit sum over neighbours.

3. Latent NameError in the no-crime-history ablation
   dataset.py called torch.zeros_like while importing only Tensor, so
   exclude_crime_history=True raised NameError the first time it ran.

Also:
- deep_baselines.py main() now builds and pre-flights all four models before
  any training: one real batch forward, shape assertions against the target,
  finite loss, backward, nonzero-gradient count. STZINB's bug surfaced only
  after three other baselines had trained, and because the exception escaped
  main() their results were never written. A shape bug now costs seconds.
- Checkpoint discovery: the Priority-2 fallback in evaluate_trained.py and
  run_conformal_evaluation.py still globbed bare run_*, re-introducing the
  contamination Priority-1 fixes. It fires exactly when no untagged run exists
  for the requested city -- i.e. mid-regeneration, when ablations may be the
  only dirs present -- and returned run_nyc_* for a Chicago request.
- run_conformal_evaluation.py's last-resort single-checkpoint fallback silently
  collapsed the multi-seed EMOS ensemble to one model while still labelling the
  result an ensemble CRPS. It now warns loudly and names the file.

Tests: 315 -> 319 + 9 new (dtype contract, value preservation, ablation path,
einsum vs explicit sum, S==W silent-corruption control, all four baselines
forward/backward, real discovery functions incl. the fallback path).
Every new test verified to fail against the pre-fix code.

+++ src/civicsafe/models/dataset.py
@@ -14,6 +14,7 @@ from __future__ import annotations
 
 import logging
 
+import torch
 from torch import Tensor
 from torch.utils.data import Dataset
 
@@ -81,14 +82,30 @@ class CrimeWindowDataset(Dataset):  # type: ignore[type-arg]
         t = self.valid_targets[idx]
         w = self.window_size
 
-        input_counts = self.counts[:, t - w : t, :]
+        # Counts arrive as int64 (panel.py builds them with dtype=torch.long,
+        # correctly -- crime counts are integers). Emit float32 instead, because
+        # every consumer is doing float arithmetic on them and torch will not
+        # infer a float output dtype from an integer input:
+        #
+        #   input_counts.mean(dim=1)
+        #   RuntimeError: mean(): could not infer output dtype.
+        #                 Input dtype must be either a floating point ... Got: Long
+        #
+        # The main model never hit this because it reaches the counts through
+        # log1p, which promotes to float32 on its own. Baselines that average or
+        # regress on the raw window hit it immediately. Casting here fixes every
+        # consumer at once rather than sprinkling .float() at each call site, and
+        # it is provably neutral for the main path: log1p(long) and
+        # log1p(float32) are bit-identical (verified over random counts to 800).
+        # float32 also halves the memory of these windows versus int64.
+        input_counts = self.counts[:, t - w : t, :].float()
         if self.exclude_crime_history:
             input_counts = torch.zeros_like(input_counts)
 
         res = {
-            "input_counts": input_counts,  # (S, W, C)
+            "input_counts": input_counts,  # (S, W, C) float32
             "input_features": self.features[:, t - w : t, :],  # (S, W, F)
-            "target_counts": self.counts[:, t, :],  # (S, C)
+            "target_counts": self.counts[:, t, :].float(),  # (S, C) float32
         }
         if self.groups is not None:
             res["groups"] = self.groups  # (S,)
